requirements.txt
google-generativeai
pytest
typing
python-dotenv
coverage
black
flake8
mypy